agregado talleres cursos

+++ app/Http/Controllers/CursoController.php
@@ -7,6 +7,7 @@ use App\Nivel;
 use App\Materia;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CursoController extends Controller
 {
@@ -17,7 +18,7 @@ class CursoController extends Controller
      */
     public function index()
     {
-
+        Gate::authorize('haveaccess', 'curso.index');
                
         $cursos= Curso::orderBy('id','Asc')->paginate(5);
        
@@ -31,6 +32,7 @@ class CursoController extends Controller
      */
     public function create()
     {
+        Gate::authorize('haveaccess', 'curso.create');
         $nivels=Nivel::get();
        
       
@@ -45,9 +47,7 @@ class CursoController extends Controller
      */
     public function store(Request $request)
     {
-       //return $request->all();
-
-        // dd($request);
+        Gate::authorize('haveaccess', 'curso.store');
         $request->validate([
 
             'nombre'      => 'required|string|max:60',
@@ -79,6 +79,7 @@ class CursoController extends Controller
      */
     public function show(Curso $curso)
     {
+        Gate::authorize('haveaccess', 'curso.show');
         return view ('Cursos.showc',['curso'=>$curso]);
     }
 
@@ -90,6 +91,7 @@ class CursoController extends Controller
      */
     public function edit(Curso $curso)
     {
+        Gate::authorize('haveaccess', 'curso.edit');
 
 
        $nivels=Nivel::get();
@@ -108,6 +110,7 @@ class CursoController extends Controller
      */
     public function update(Request $request, Curso $curso)
     {
+        Gate::authorize('haveaccess', 'curso.update');
         
         $request->validate([
 
@@ -135,6 +138,7 @@ class CursoController extends Controller
      */
     public function destroy(Curso $curso)
     {
+        Gate::authorize('haveaccess', 'curso.destroy');
         $curso= Curso::find($curso->id);
         $curso->delete();
